8//
-Fixed a bug where it would not wait for the modal to be built before trying to open it.
-Fixed some typos
-Added a tad bit of defensive programming so one cannot request more characters than available.
-Fixed a flaw when it could enter an infinite loop trying to generate a location while no valid location exist
(eg : requesting 100 individual cards of "unknown". There are dupplicates in those 100 cards so the program could never reach 100 individual cards.)
-Fixed a dum dum on my part in the requestCharacters function, was filling the notAllowedPageIndexPairs twice on each pass.
-Also fixed comments (related to my dum dum brain) stating I was passing an array as parameter when I was in fact passing a pointer to that array.
-Added a function to fetch data and changed Episode and Character initialize methods to use it instead.

Note : unexcpected behavior can occur on large fetches (eg : 826 cards of "all")
If a fetch fail it will try again a set amount of times, but after that the code continue and might crash, (couldn't make it happen but that's a possibility)
Also since I do not handle the fetch of the pictures (I only put the appropriate href url in the <img> tag, the browser then does it's thing) , large fetches of pictures could fail due to "Too many requests"
Not implemented but a solution could be to load only a few rows and add a button "load more"

On branch main
Your branch is behind 'origin/main' by 1 commit, and can be fast-forwarded.
  (use "git pull" to update your local branch)

Changes to be committed:
	new file:   public/custom.css
	modified:   public/index.html
	modified:   public/script.js
	modified:   src/styles.css